+++ src/app.module.ts
@@ -7,7 +7,7 @@ import authConfig from './config/auth.config';
 
 import { UsersModule } from './users/users.module';
 import { IamModule } from './iam/iam.module';
-import { SubappsModule } from './subapps/resources/subapps.module';
+import { SubappsModule } from './subapps/subapps.module';
 import { MyfreeinvModule } from './subapps/myfreeinv/myfreeinv.module';
 import { BizlinksfreeModule } from './subapps/bizlinksfree/bizlinksfree.module';
 import { MycuttingboardModule } from './subapps/mycuttingboard/mycuttingboard.module';
