+++ src/subapps/subapps.module.ts
@@ -1,16 +1,16 @@
 import { Module } from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config';
-import { MyfreeinvModule } from '../myfreeinv/myfreeinv.module';
-import { BizlinksfreeModule } from '../bizlinksfree/bizlinksfree.module';
+import { MyfreeinvModule } from './myfreeinv/myfreeinv.module';
+import { BizlinksfreeModule } from './bizlinksfree/bizlinksfree.module';
 import { TypeOrmModule } from '@nestjs/typeorm';
 
 import { SubappsService } from './subapps.service';
 import { SubappsController } from './subapps.controller';
 
-import { UserSubappAccess } from './entities/userSubappAccess.entity';
+import { UserSubappAccess } from './resources/entities/userSubappAccess.entity';
 
 import authConfig from 'src/config/auth.config';
-import { MycuttingboardModule } from '../mycuttingboard/mycuttingboard.module';
+import { MycuttingboardModule } from './mycuttingboard/mycuttingboard.module';
 
 @Module({
   imports: [
